refactor: replaced task-related classes to the directory task
