+++ tasks/gabrielsebrao/ToDo/app/src/main/java/com/example/todo/adapter/ArchivedTaskAdapter.kt
@@ -10,8 +10,7 @@ import android.widget.TextView
 import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
 import com.example.todo.R
-import com.example.todo.TaskActionListener
-import com.example.todo.room.Task
+import com.example.todo.task.Task
 
 class ArchivedTaskAdapter(
     val taskList: MutableList<Task>
