+++ tasks/gabrielsebrao/ToDo/app/src/main/java/com/example/todo/room/DataBase.kt
@@ -4,6 +4,8 @@ import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
+import com.example.todo.task.Task
+import com.example.todo.task.TaskDao
 
 @Database(entities = [Task::class], version = 1)
 abstract class DataBase : RoomDatabase() {
