+++ tasks/gabrielsebrao/ToDo/app/src/main/java/com/example/todo/task/TaskDao.kt
@@ -1,9 +1,10 @@
-package com.example.todo.room
+package com.example.todo.task
 
 import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.Query
+import com.example.todo.task.Task
 import io.reactivex.rxjava3.core.Completable
 
 @Dao
