+++ tasks/gabrielsebrao/ToDo/app/src/main/java/com/example/todo/task/TaskSingleton.kt
@@ -1,6 +1,6 @@
-package com.example.todo
+package com.example.todo.task
 
-import com.example.todo.room.Task
+import com.example.todo.task.Task
 
 object TaskSingleton {
